+++ HISTORY.rst
@@ -2,8 +2,17 @@
 History
 =======
 
-0.1.0 (Unreleased)
+1.1.0 (Unreleased)
 ------------------
+
+* Changed address mixin to use new Point custom type with GeoJSON support. (ericof)
+* Add unit tests to address mixin and improve workflow database unit tests. (ericof)
+* Improve address mixin: metadata for colander alchemy and schema validators. (rudaporto)
+* Change travis run flake8 on tests folder. (rudaporto)
+
+
+0.1.0
+-----
 * overhauls workflow (jsbueno)
 * Implements timeout_cache (jsbueno)
 * Add helpers to deal with coordinates entry as a GeoJSON (ericof)